refactor: rename file for cloud formation client
