+++ packages/core/src/lambda/commands/deleteCloudFormation.ts
@@ -7,7 +7,7 @@ import * as nls from 'vscode-nls'
 const localize = nls.loadMessageBundle()
 
 import * as vscode from 'vscode'
-import { CloudFormationClient } from '../../shared/clients/cloudFormationClient'
+import { CloudFormationClient } from '../../shared/clients/cloudFormation'
 
 import * as localizedText from '../../shared/localizedText'
 import { getLogger, Logger } from '../../shared/logger/logger'
