+++ packages/core/src/shared/sam/deploy.ts
@@ -8,7 +8,7 @@ import { AWSTreeNodeBase } from '../treeview/nodes/awsTreeNodeBase'
 import { TreeNode, isTreeNode } from '../treeview/resourceTreeDataProvider'
 import globals from '../../shared/extensionGlobals'
 import { ToolkitError } from '../../shared/errors'
-import { CloudFormationClient } from '../clients/cloudFormationClient'
+import { CloudFormationClient } from '../clients/cloudFormation'
 import { S3Client } from '../clients/s3'
 import { samDeployUrl } from '../constants'
 import { getSpawnEnv } from '../env/resolveEnv'
