+++ packages/core/src/shared/sam/sync.ts
@@ -10,7 +10,7 @@ import * as path from 'path'
 import * as localizedText from '../localizedText'
 import { S3Client } from '../clients/s3'
 import { DataQuickPickItem, createMultiPick, createQuickPick } from '../ui/pickerPrompter'
-import { CloudFormationClient } from '../clients/cloudFormationClient'
+import { CloudFormationClient } from '../clients/cloudFormation'
 import * as CloudFormation from '../cloudformation/cloudformation'
 import { DefaultEcrClient } from '../clients/ecrClient'
 import { createRegionPrompter } from '../ui/common/region'
