+++ packages/core/src/test/shared/sam/deploy.test.ts
@@ -13,13 +13,13 @@ import { samconfigCompleteData, samconfigInvalidData, validTemplateData } from '
 import * as SamUtilsModule from '../../../shared/sam/utils'
 import assert from 'assert'
 import { getTestWindow } from '../vscode/window'
-import { CloudFormationClient } from '../../../shared/clients/cloudFormationClient'
+import { CloudFormationClient } from '../../../shared/clients/cloudFormation'
 import { intoCollection } from '../../../shared/utilities/collectionUtils'
 import { clickBackButton, createPromptHandler, PrompterTester } from '../wizards/prompterTester'
 import { RegionNode } from '../../../awsexplorer/regionNode'
 import { createTestRegionProvider } from '../regions/testUtil'
 import { S3Client } from '../../../shared/clients/s3'
-import * as CloudFormationClientModule from '../../../shared/clients/cloudFormationClient'
+import * as CloudFormationClientModule from '../../../shared/clients/cloudFormation'
 import * as S3ClientModule from '../../../shared/clients/s3'
 import * as ProcessUtilsModule from '../../../shared/utilities/processUtils'
 import * as ProcessTerminalModule from '../../../shared/sam/processTerminal'
