+++ packages/core/src/test/shared/ui/sam/stackPrompter.test.ts
@@ -10,7 +10,7 @@ import * as vscode from 'vscode'
 import * as AwsConsoleModule from '../../../../shared/awsConsole'
 import * as SamUtilsModule from '../../../../shared/sam/utils'
 import * as ButtonsModule from '../../../../shared/ui/buttons'
-import { CloudFormationClient } from '../../../../shared/clients/cloudFormationClient'
+import { CloudFormationClient } from '../../../../shared/clients/cloudFormation'
 import { samSyncUrl } from '../../../../shared/constants'
 import { createStackPrompter } from '../../../../shared/ui/sam/stackPrompter'
 import { intoCollection } from '../../../../shared/utilities/collectionUtils'
